fix(messageProviderPact.ts): close the http server instance started by the Message Provider Verifier
When running MessageProvider tests using Jest the test pack never quits. This has been narrowed down
to a http server that is started when calling the verify() method but is never closed. A small
change to ensure that the server gets shut down correctly after the verification has been completed.