feat: Automatically set `changeOrigin: true` for non-local verifications  (#319)
* feat(verification): automatically set changeOrigin:true for non-local verification

When the provider verification process is run against a remote provider, the origin
header can mismatch the actual remote proxied process.

This PR defaults to change the origin of the host header to the target URL, unless
specifically overridden by the user. This should remove a class of confusing issues
for new users.

See #280, #281 and #282 for background. This specific issue has come up on Slack
a number of times, and hence fixing before it appears as a GH issue.

* chore(build): allow ipv6 networking in travis

* chore(build): allow ipv6 networking in travis

* chore(build): allow ipv6 networking in travis