fix(logger): Eliminate memory leak for sequential run in jest
`.pipe()` call attaches events listeners to the writeable stream
which in case of stdout and sequential jest tests causes memory
leak as those listeners are added for every test file.

Since stdout stream is special, for example it never ends while
the process is runing, we can use simplified piping by
listening to the `data` event of the readable stream and
redirect data to stdout. This way no listeners are leaking.