chore: update to install binary on 'postinstall'
Scenario: we always run all examples against the current, in-flight build,
to ensure they are both relevant and correct. To avoid confusion for users,
we install the real version in package.json (rather than use a relative
import in the code) and `npm link` in builds - this means people can
copy/paste, and we are also testing the externally facing API.

The issue is, that this requires an `npm i` first before we link, which will attempt
to download the version of Pact from npmjs, which in the `install` phase will in turn
attempt to download a native binary from github for that version/os/arch - and this won't
exist for the case when we introduce a new version of Node.

The fix for these cases is to allow our builds to fallback to a Rust cargo build, but
doing the binary download on the `install` phase means the source code is not installed,
and the Cargo.yaml file and associated code doesn't exist.

Moving to `postinstall` means this should be fixed, and installing neon into the examples
means the fallback build can take place (on CI).