feat: support promises in filters + state handlers
Adds back the proxy infrastructure to allow native promises
in the request filters and provider state handlers.

BREAKING CHANGE: the signature of state handlers has been updated to
accept either a single function with parameters, or an object that
can specify optional teardown and setup functions that run on the
different state phases.

BREAKING CHANGE: callbackTimeout is now timeout