Lviana menlosecurity patch 2 (#701)
fix: avoid exception when printing request.body

When a request is not received we are supposed to get an error message saying "The following request was expected but not received", with the request printed immediately afterwards. However, if the request has a JSON body, we get an exception saying "TypeError: request.body.substr is not a function". The problem is that we need to first convert the JSON body to a string before we can apply the substr method to it (and similarly for the length property). This PR fixes that.