refactor: pass all options down to server
Options which are added to the server will now be automatically passed
down without a need for modifications in this library, other than
exposing the new options in the types, if that is desirable.

This allows consumers to pass down the monkeypatch option when creating
the server, but that option is intentionally excluded from the
PactOptions type, as this option is advanced and its use is discouraged.