chore: prevent relative import paths
Some IDEs (such as VS Code) are aware of the `baseUrl'
property, and will auto-complete imports using the base
URL as a prefix in the import, instead of a relative URL.

This creates problems for consumers of the library, as the path
is no longer relative to the top level project base (I think).

Even if this is untrue, it's best to stick with a single,
consistent approach.

As an example, see a20ad80 which had to correct one such
problem.

See also #974