fix: allow hooks to run only once per interaction
Current behaviour will run the beforeEach and afterEach hooks multiple
times if there are several provider states defined in an interaction.
This is due to the way the hook are run on every call to "/_pactSetup"
via the proxy.

This change will ensure each of those hooks is run only once per
interaction, regardless of how many provider states are defined in that
interaction.