chore(ci): deduplicate CI runs
We don't need to run CI on every push to any branch _and_ PRs targetting
master. This change ensures that the following are targetted:

- New commits to `master`
- PRs targetting `master`

Specifically, commits to new branches are _not_ tested unless they are
part of a PR.

Lastly, removed the `9.x.x` mention as this does not appear to be a
branch that exists.

Signed-off-by: JP-Ellis <josh@jpellis.me>