chore: rewrite test workflow
The motivation for this is improve the speed at which the CI tasks
complete in the Pact-JS repository. This hopefully achieves this by
having more parallel builds which are each smaller and cached.

I have also reduced the reliance on the scripts as it appeared to me
that they duplicate a number of steps. These duplications are justified
when running locally to ensure all previous steps are run, but in CI it
likely results in longer runs thhan necessary.

Signed-off-by: JP-Ellis <josh@jpellis.me>